fixed: nrf_service (move SendNFDiscoveryXXX from udm_service to nrf_service)
