+++ internal/sbi/consumer/udm_service.go
@@ -1,11 +1,11 @@
 package consumer
 
 import (
-	"net/http"
+	// "net/http"
 	"sync"
 
 	// "github.com/antihax/optional"
-	"github.com/mohae/deepcopy"
+	// "github.com/mohae/deepcopy"
 	"github.com/pkg/errors"
 
 	"github.com/free5gc/openapi"
@@ -69,115 +69,6 @@ func (s *nudmService) getUEContextManagementClient(uri string) *Nudm_UEContextMa
 	defer s.UEContextManagementMu.Unlock()
 	s.UEContextManagementClients[uri] = client
 	return client
-}
-
-func (s *nudmService) SendNFDiscoveryUDM() (*models.ProblemDetails, error) {
-	ctx, pd, err := smf_context.GetSelf().GetTokenCtx(models.ServiceName_NNRF_DISC, models.NfType_NRF)
-	if err != nil {
-		return pd, err
-	}
-
-	smfContext := s.consumer.smf.Context()
-
-	// Check data
-	result, httpResp, localErr := s.consumer.nnrfService.NFDiscoveryUDM(ctx)
-
-	if localErr == nil {
-		smfContext.UDMProfile = result.NfInstances[0]
-
-		for _, service := range *smfContext.UDMProfile.NfServices {
-			if service.ServiceName == models.ServiceName_NUDM_SDM {
-				smfContext.SubscriberDataManagementClient = s.consumer.getSubscribeDataManagementClient(service.ApiPrefix)
-			}
-		}
-
-		if smfContext.SubscriberDataManagementClient == nil {
-			logger.ConsumerLog.Warnln("sdm client failed")
-		}
-	} else if httpResp != nil {
-		defer func() {
-			if resCloseErr := httpResp.Body.Close(); resCloseErr != nil {
-				logger.ConsumerLog.Errorf("SearchNFInstances response body cannot close: %+v", resCloseErr)
-			}
-		}()
-		logger.ConsumerLog.Warnln("handler returned wrong status code ", httpResp.Status)
-		if httpResp.Status != localErr.Error() {
-			return nil, localErr
-		}
-		problem := localErr.(openapi.GenericOpenAPIError).Model().(models.ProblemDetails)
-		return &problem, nil
-	} else {
-		return nil, openapi.ReportError("server no response")
-	}
-	return nil, nil
-}
-
-func (s *nudmService) SendNFDiscoveryPCF() (problemDetails *models.ProblemDetails, err error) {
-	ctx, pd, err := smf_context.GetSelf().GetTokenCtx(models.ServiceName_NNRF_DISC, models.NfType_NRF)
-	if err != nil {
-		return pd, err
-	}
-
-	// Check data
-	result, httpResp, localErr := s.consumer.nnrfService.NFDiscoveryPCF(ctx)
-
-	if localErr == nil {
-		logger.ConsumerLog.Traceln(result.NfInstances)
-	} else if httpResp != nil {
-		defer func() {
-			if resCloseErr := httpResp.Body.Close(); resCloseErr != nil {
-				logger.ConsumerLog.Errorf("SearchNFInstances response body cannot close: %+v", resCloseErr)
-			}
-		}()
-		logger.ConsumerLog.Warnln("handler returned wrong status code ", httpResp.Status)
-		if httpResp.Status != localErr.Error() {
-			err = localErr
-			return problemDetails, err
-		}
-		problem := localErr.(openapi.GenericOpenAPIError).Model().(models.ProblemDetails)
-		problemDetails = &problem
-	} else {
-		err = openapi.ReportError("server no response")
-	}
-
-	return problemDetails, err
-}
-
-func (s *nudmService) SendNFDiscoveryServingAMF(smContext *smf_context.SMContext) (*models.ProblemDetails, error) {
-	ctx, pd, err := smf_context.GetSelf().GetTokenCtx(models.ServiceName_NNRF_DISC, models.NfType_NRF)
-	if err != nil {
-		return pd, err
-	}
-
-	// Check data
-	result, httpResp, localErr := s.consumer.nnrfService.NFDiscoveryAMF(smContext, ctx)
-
-	if localErr == nil {
-		if result.NfInstances == nil {
-			if status := httpResp.StatusCode; status != http.StatusOK {
-				logger.ConsumerLog.Warnln("handler returned wrong status code", status)
-			}
-			logger.ConsumerLog.Warnln("NfInstances is nil")
-			return nil, openapi.ReportError("NfInstances is nil")
-		}
-		logger.ConsumerLog.Info("SendNFDiscoveryServingAMF ok")
-		smContext.AMFProfile = deepcopy.Copy(result.NfInstances[0]).(models.NfProfile)
-	} else if httpResp != nil {
-		defer func() {
-			if resCloseErr := httpResp; resCloseErr != nil {
-				logger.ConsumerLog.Errorf("SearchNFInstances response body cannot close: %+v", resCloseErr)
-			}
-		}()
-		if httpResp.Status != localErr.Error() {
-			return nil, localErr
-		}
-		problem := localErr.(openapi.GenericOpenAPIError).Model().(models.ProblemDetails)
-		return &problem, nil
-	} else {
-		return nil, openapi.ReportError("server no response")
-	}
-
-	return nil, nil
 }
 
 func (s *nudmService) UeCmRegistration(smCtx *smf_context.SMContext) (
